fixed JSON file generation
fix: JSON tweet data is now working as intended

The JSON file with the tweet data for the entire [non-deleted] list of tweets is now saving correctly to be read back into the master notebook.

Created a copy of the file to avoid having to wait for the twitter API rate-limit if original is erased or altered during wrangling or analysis steps.